Moved updater into commands
