--- org.eclipse.ice.commands.notification/.gitignore
@@ -1,2 +1,0 @@
-/target/
-/bin/
