+++ org.eclipse.ice.commands/pom.xml
@@ -35,5 +35,15 @@
         	<artifactId>sshd-sftp</artifactId>
         	<version>2.3.0</version>
         </dependency>
+        <dependency>
+      		<groupId>com.sun.mail</groupId>
+      		<artifactId>javax.mail</artifactId>
+      		<version>1.6.2</version>
+    	</dependency>
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
+      		<version>4.5.10</version>
+        </dependency>
    	</dependencies>
 </project>
