+++ org.eclipse.ice.commands/src/test/java/org/eclipse/ice/tests/commands/HTTPHandlerTest.java
@@ -9,11 +9,12 @@
  *   Initial API and implementation and/or initial documentation - 
  *   Jay Jay Billings, Joe Osborn
  *******************************************************************************/
-package org.eclipse.ice.commands.notification;
+package org.eclipse.ice.tests.commands;
 
 import java.io.IOException;
 
 import org.eclipse.ice.commands.CommandStatus;
+import org.eclipse.ice.commands.HTTPCommandUpdateHandler;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
